+++ app/src/main/java/org/hyperledger/besu/controller/BesuControllerBuilder.java
@@ -58,6 +58,8 @@ import org.hyperledger.besu.ethereum.eth.manager.MonitoredExecutors;
 import org.hyperledger.besu.ethereum.eth.manager.peertask.PeerTaskExecutor;
 import org.hyperledger.besu.ethereum.eth.manager.peertask.PeerTaskRequestSender;
 import org.hyperledger.besu.ethereum.eth.manager.snap.SnapProtocolManager;
+import org.hyperledger.besu.ethereum.eth.peervalidation.ClassicForkPeerValidator;
+import org.hyperledger.besu.ethereum.eth.peervalidation.DaoForkPeerValidator;
 import org.hyperledger.besu.ethereum.eth.peervalidation.PeerValidator;
 import org.hyperledger.besu.ethereum.eth.peervalidation.RequiredBlocksPeerValidator;
 import org.hyperledger.besu.ethereum.eth.sync.DefaultSynchronizer;
@@ -1438,6 +1440,21 @@ public abstract class BesuControllerBuilder implements MiningConfigurationOverri
   protected List<PeerValidator> createPeerValidators(
       final ProtocolSchedule protocolSchedule, final PeerTaskExecutor peerTaskExecutor) {
     final List<PeerValidator> validators = new ArrayList<>();
+
+    final OptionalLong daoBlock = genesisConfigOptions.getDaoForkBlock();
+    if (daoBlock.isPresent()) {
+      // Setup dao validator
+      validators.add(
+          new DaoForkPeerValidator(protocolSchedule, peerTaskExecutor, daoBlock.getAsLong()));
+    }
+
+    final OptionalLong classicBlock = genesisConfigOptions.getClassicForkBlock();
+    // setup classic validator
+    if (classicBlock.isPresent()) {
+      validators.add(
+          new ClassicForkPeerValidator(
+              protocolSchedule, peerTaskExecutor, classicBlock.getAsLong()));
+    }
 
     for (final Map.Entry<Long, Hash> requiredBlock : requiredBlocks.entrySet()) {
       validators.add(