+++ app/src/main/java/org/hyperledger/besu/controller/MainnetBesuControllerBuilder.java
@@ -36,7 +36,7 @@ import java.util.Optional;
 /** The Mainnet besu controller builder. */
 public class MainnetBesuControllerBuilder extends BesuControllerBuilder {
 
-  private final EpochCalculator epochCalculator = new EpochCalculator.DefaultEpochCalculator();
+  private EpochCalculator epochCalculator = new EpochCalculator.DefaultEpochCalculator();
 
   /** Default constructor. */
   public MainnetBesuControllerBuilder() {}
@@ -102,6 +102,9 @@ public class MainnetBesuControllerBuilder extends BesuControllerBuilder {
 
   @Override
   protected void prepForBuild() {
-    // No special preparation needed for mainnet
+    genesisConfigOptions
+        .getThanosBlockNumber()
+        .ifPresent(
+            activationBlock -> epochCalculator = new EpochCalculator.Ecip1099EpochCalculator());
   }
 }