+++ app/src/test/java/org/hyperledger/besu/cli/NetworkDeprecationMessageTest.java
@@ -26,7 +26,16 @@ class NetworkDeprecationMessageTest {
   @ParameterizedTest
   @EnumSource(
       value = NetworkDefinition.class,
-      names = {"MAINNET", "SEPOLIA", "DEV", "LUKSO", "EPHEMERY", "HOODI"})
+      names = {"HOLESKY"})
+  void shouldGenerateDeprecationMessageForDeprecatedNetworks(final NetworkDefinition network) {
+    assertThat(NetworkDeprecationMessage.generate(network))
+        .contains(network.normalize() + " is deprecated");
+  }
+
+  @ParameterizedTest
+  @EnumSource(
+      value = NetworkDefinition.class,
+      names = {"MAINNET", "SEPOLIA", "DEV", "LUKSO", "EPHEMERY", "HOODI", "CLASSIC", "MORDOR"})
   void shouldThrowErrorForNonDeprecatedNetworks(final NetworkDefinition network) {
     assertThatThrownBy(() -> NetworkDeprecationMessage.generate(network))
         .isInstanceOf(AssertionError.class);